refactor(app): move the session log tab worker into ainb-app
fleet::session_log reads a session's notification history off the render
path. It was the last fleet module held in ainb-core, only because its row
type and filter lived in session_tabs. LogRow, log_rows and log_detail move
to ainb_app::components::session_tabs, session_log follows, and the core
fleet shim is gone.
