+++ ainb-tui/crates/ainb-app/src/components/mod.rs
@@ -9,6 +9,7 @@ pub mod log_parser;
 pub mod log_reader;
 pub mod log_writer;
 pub mod onboarding;
+pub mod session_tabs;
 pub mod setup_menu;
 pub mod skill_manager_screen;
 pub mod skills;
