+++ ainb-tui/crates/ainb-app/src/fleet/session_log.rs
