--- ainb-tui/crates/ainb-core/src/fleet/mod.rs
@@ -1,7 +1,0 @@
-// ABOUTME: Terminal-side fleet pieces. The fleet service layer lives in
-// `ainb-app`; this module re-exports it and adds the session log tab worker, which still
-// depends on a screen component.
-
-pub use ainb_app::fleet::*;
-
-pub mod session_log;
